pyproject.toml
[tool.poetry]
name = "django-project"
version = "0.1.0"
description = ""
authors = ["chanhong-ji <jichanhong24@gmail.com>"]
readme = "README.md"
packages = [{include = "django_project"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.6"
mysqlclient = "^2.1.1"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
